♻️ Gas limit (uint32 -> uint256)

+++ contracts/L2_NovaRegistry.sol
@@ -70,7 +70,7 @@ contract L2_NovaRegistry is ReentrancyGuard, OVM_CrossDomainEnabled {
     /// @dev Maps execHashes to the calldata associated with the request.
     mapping(bytes32 => bytes) private requestCalldatas;
     /// @dev Maps execHashes to the gas limit a bot should use to execute the request.
-    mapping(bytes32 => uint32) private requestGasLimits;
+    mapping(bytes32 => uint256) private requestGasLimits;
     /// @dev Maps execHashes to the gas price a bot must use to execute the request.
     mapping(bytes32 => uint256) private requestGasPrices;
     /// @dev Maps execHashes to the 'bounty' tokens a bot will recieve for executing the request.
@@ -99,7 +99,7 @@ contract L2_NovaRegistry is ReentrancyGuard, OVM_CrossDomainEnabled {
             // General request data:
             address strategy,
             bytes memory l1calldata,
-            uint32 gasLimit,
+            uint256 gasLimit,
             uint256 gasPrice,
             InputToken[] memory inputTokens,
             Bounty[] memory bounties,
@@ -132,7 +132,7 @@ contract L2_NovaRegistry is ReentrancyGuard, OVM_CrossDomainEnabled {
     function requestExec(
         address strategy,
         bytes calldata l1calldata,
-        uint32 gasLimit,
+        uint256 gasLimit,
         uint256 gasPrice,
         InputToken[] calldata inputTokens,
         Bounty[] calldata bounties
@@ -183,7 +183,7 @@ contract L2_NovaRegistry is ReentrancyGuard, OVM_CrossDomainEnabled {
     function requestExecWithTimeout(
         address strategy,
         bytes calldata l1calldata,
-        uint32 gasLimit,
+        uint256 gasLimit,
         uint256 gasPrice,
         InputToken[] calldata inputTokens,
         Bounty[] calldata bounties,
@@ -280,7 +280,7 @@ contract L2_NovaRegistry is ReentrancyGuard, OVM_CrossDomainEnabled {
             )
         );
 
-        uint32 gasLimit = requestGasLimits[execHash];
+        uint256 gasLimit = requestGasLimits[execHash];
 
         // Fill out data for the resubmitted request.
         requestStrategies[newExecHash] = requestStrategies[execHash];
